// webui/src/proto/orchestrator_account.ts
export interface UpdateAccountAuthRequest {
  session_token: string;
  access_token: string;
  activation_channel?: string | undefined;
}